Move objects of gateway to lib
