--- crates/papyrus_gateway/src/objects/mod.rs
@@ -1,10 +1,0 @@
-mod block;
-mod state;
-mod transaction;
-
-pub use block::{Block, BlockHeader};
-pub use state::{ContractClass, StateUpdate, ThinStateDiff};
-pub use transaction::{
-    Event, Transaction, TransactionOutput, TransactionReceipt, TransactionReceiptWithStatus,
-    TransactionStatus, TransactionWithType, Transactions,
-};
